+++ src/lib/rxfm/components/index.ts
@@ -1,4 +1,5 @@
 export * from './component';
+export * from './chainable-creator';
 export * from './html';
 export * from './svg';
 export * from './add-to-view';